requirements.txt
aiogram
tortoise-orm[asyncpg]
aerich
python-dotenv
